improved structure in channels pages
