--- frontend/ui/src/pages/Channels/ChannelsMainPage/SourceDescription/index.module.scss
@@ -1,100 +1,0 @@
-@import '../../../../assets/scss/fonts';
-@import '../../../../assets/scss/colors';
-
-.channelCard {
-  display: flex;
-  box-sizing: border-box;
-  height: 100px;
-  width: 300px;
-  border: 1px solid var(--color-light-gray);
-  border-radius: 8px;
-  background-color: white;
-  margin-top: 24px;
-}
-
-.channelLogo {
-  display: inline-flex;
-  height: 40px;
-  width: 40px;
-  margin: 16px 8px 42px 16px;
-  svg {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.channelTitle {
-  display: inline-flex;
-  height: 24px;
-  width: 190px;
-  color: #212428;
-  font-size: 16px;
-  font-weight: bold;
-  letter-spacing: 0;
-  line-height: 24px;
-  margin: 16px 20px 2px 0;
-}
-
-.channelText {
-  display: inline-flex;
-  height: 48px;
-  width: 190px;
-  color: var(--color-text-gray);
-  font-size: 16px;
-  letter-spacing: 0;
-  line-height: 24px;
-}
-
-.channelButton {
-  margin: 20px 0px 8px 0px;
-}
-
-.addChannelButton {
-  position: relative;
-  background: white;
-  width: 40px;
-  height: 40px;
-  margin-left: 7px;
-  margin-top: 32px;
-  transition: 0.2s ease-in-out all;
-  border: none;
-  padding: 0;
-  outline: none;
-  cursor: pointer;
-}
-
-.channelButtonIcon {
-  @include font-m;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 30px;
-  height: 30px;
-  margin-right: 16px;
-  padding: 0px;
-  border-radius: 50%;
-  color: var(--color-text-gray);
-  cursor: pointer;
-
-  svg {
-    height: 30px;
-    width: 30px;
-    margin: 0;
-    path {
-      fill: var(--color-text-gray);
-    }
-  }
-  &:hover {
-    color: var(--color-airy-blue);
-    background-color: var(--color-background-blue);
-    box-shadow: 0px 0px 0px 3px var(--color-background-blue);
-  }
-}
-
-.requirementsDialog {
-  z-index: 1;
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
