--- frontend/ui/src/pages/Channels/ChannelsMainPage/Sources/TwilioSmsConnect.module.scss
@@ -1,37 +1,0 @@
-@import '../../../../assets/scss/fonts';
-@import '../../../../assets/scss/colors';
-
-.wrapper {
-  background: white;
-  display: block;
-  border-radius: 10px;
-  padding-left: 96px;
-  padding-top: 88px;
-  width: 100%;
-  overflow-y: auto;
-  padding: 32px;
-  margin: 88px 1.5em 0 100px;
-  min-height: calc(100vh - 170px);
-}
-
-.headline {
-  @include font-xl;
-  font-weight: bold;
-  margin-bottom: 8px;
-}
-
-.backButton {
-  margin-bottom: 16px;
-  cursor: pointer;
-  text-decoration: none;
-  &:hover {
-    text-decoration: underline;
-  }
-}
-
-.backIcon {
-  path {
-    fill: var(--color-airy-blue);
-  }
-  margin-right: 8px;
-}
