--- frontend/ui/src/pages/Channels/ChannelsMainPage/index.module.scss
@@ -1,31 +1,0 @@
-@import '../../../assets/scss/fonts';
-@import '../../../assets/scss/colors';
-
-.channelsHeadline {
-  @include font-base;
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-  color: var(--color-text-contrast);
-  font-size: 39px;
-  font-weight: 900;
-  letter-spacing: 0;
-  margin-bottom: 14px;
-}
-
-.channelsHeadlineText {
-  @include font-xl;
-  font-weight: 900;
-}
-
-.channelsChoice {
-  list-style: none;
-  color: var(--color-text-gray);
-  padding-bottom: 20px;
-}
-
-.wrapper {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 1em;
-}
