+++ frontend/ui/src/pages/Channels/ConnectedChannelsBySourceCard/index.module.scss
@@ -1,0 +1,156 @@
+@import 'assets/scss/fonts.scss';
+@import 'assets/scss/colors.scss';
+
+.connectedContainer {
+  display: flex;
+  flex-direction: column;
+  min-width: 514px;
+  height: 120px;
+  margin-bottom: 10px;
+}
+.connectedSum {
+  margin-left: 16px;
+  color: var(--color-text-gray);
+}
+.connectedChannelBox {
+  display: flex;
+  box-sizing: border-box;
+  width: 550px;
+  min-height: 100px;
+  overflow: hidden;
+  border: 1px solid var(--color-light-gray);
+  border-radius: 8px;
+  background-color: #ffffff;
+  margin-left: 12px;
+  position: relative;
+
+  &:hover {
+    background-color: var(--color-background-blue);
+    box-shadow: -0.5px 2px 3px 0px var(--color-light-gray);
+    cursor: pointer;
+    button {
+      background-color: var(--color-background-blue);
+      cursor: pointer;
+    }
+  }
+}
+.connectedChannel {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: column;
+  justify-content: space-evenly;
+  white-space: nowrap;
+  min-width: 400px;
+  min-height: 94px;
+  padding: 8px;
+}
+
+.extraPhoneInfo {
+  overflow: hidden;
+  max-width: 142px;
+  text-overflow: ellipsis;
+}
+.extraChannel {
+  display: flex;
+  min-width: 88px;
+  width: 100%;
+  min-height: 20px;
+  padding: 0 24px 0 24px;
+  cursor: pointer;
+  button {
+    outline: none;
+  }
+}
+
+.connectedChannelData {
+  @include font-base;
+  display: flex;
+  align-items: center;
+  margin: 4px;
+  border: none;
+  outline: none;
+  background-color: inherit;
+  text-align: left;
+  cursor: pointer;
+  outline: none;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.channelListEntry {
+  display: inline-flex;
+}
+.placeholderLogo {
+  height: 24px;
+  width: 24px;
+  margin-right: 4px;
+  svg {
+    height: 24px;
+    width: 24px;
+  }
+}
+.facebookImage {
+  height: 20px;
+  width: 20px;
+  margin-right: 4px;
+  border-radius: 50%;
+}
+
+.connectedChannelName {
+  min-width: 168px;
+  max-width: 200px;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
+  margin-right: 7px;
+  &:empty {
+    display: none;
+  }
+}
+
+.channelId {
+  padding-top: 10px;
+}
+
+.channelButton {
+  margin: 20px 0px 8px 0px;
+}
+.addChannelButton {
+  position: relative;
+  width: 40px;
+  height: 40px;
+  background: white;
+  margin: 32px 0px 0px 7px;
+  padding: 0;
+  transition: 0.2s ease-in-out all;
+  border: none;
+  outline: none;
+  cursor: pointer;
+}
+.channelButtonIcon {
+  @include font-m;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  width: 30px;
+  height: 30px;
+  margin-right: 16px;
+  padding: 0px;
+  border-radius: 50%;
+  color: var(--color-text-gray);
+  cursor: pointer;
+  svg {
+    height: 30px;
+    width: 30px;
+    margin: 0;
+    path {
+      fill: var(--color-text-gray);
+    }
+  }
+  &:hover {
+    color: var(--color-airy-blue);
+    background-color: var(--color-background-blue);
+  }
+}
