+++ frontend/ui/src/pages/Channels/ConnectedChannelsBySourceCard/index.tsx
@@ -1,0 +1,88 @@
+import React from 'react';
+
+import {LinkButton} from '@airyhq/components';
+import {Channel} from 'httpclient';
+import {fallbackImage} from '../../../services/image/index';
+import {ReactComponent as PlusCircleIcon} from 'assets/images/icons/plus-circle.svg';
+
+import styles from './index.module.scss';
+
+type ConnectedChannelsBySourceCardProps = {
+  source: string;
+  channels: Channel[];
+  connected: string;
+  placeholderImage?: JSX.Element;
+  isConnected: string;
+  onAddChannelClick?: () => void;
+  onMoreChannelsClick?: () => void;
+  onChannelClick?: (channel: Channel) => void;
+  onSourceInfoClick?: () => void;
+};
+
+const ConnectedChannelsBySourceCard = (props: ConnectedChannelsBySourceCardProps) => {
+  const {source, channels} = props;
+  const isPhoneNumberSource = () => {
+    return source === 'twilio.sms' || source === 'twilio.whatsapp';
+  };
+  const channelsToShow = isPhoneNumberSource() ? 2 : 4;
+  const hasExtraChannels = channels.length > channelsToShow;
+  return (
+    <>
+      {channels && channels.length > 0 && (
+        <>
+          <div className={styles.connectedContainer}>
+            <div className={styles.connectedSum}>
+              <p>
+                {channels.length} {props.connected}
+              </p>
+            </div>
+            <div className={styles.connectedChannelBox}>
+              <div className={styles.connectedChannel}>
+                {channels.slice(0, channelsToShow).map((channel: Channel) => {
+                  return (
+                    <li key={channel.sourceChannelId} className={styles.channelListEntry}>
+                      <button className={styles.connectedChannelData} onClick={() => props.onChannelClick(channel)}>
+                        {source === 'facebook' && channel.metadata?.imageUrl ? (
+                          <img
+                            src={channel.metadata?.imageUrl}
+                            alt={channel.metadata?.name}
+                            className={styles.facebookImage}
+                            onError={(event: React.SyntheticEvent<HTMLImageElement, Event>) => {
+                              fallbackImage(event, channel.source);
+                            }}
+                          />
+                        ) : (
+                          <div className={styles.placeholderLogo}>{props.placeholderImage} </div>
+                        )}
+                        <div className={styles.connectedChannelName}>{channel.metadata?.name}</div>
+                        {isPhoneNumberSource() && (
+                          <div className={styles.extraPhoneInfo}>{channel.sourceChannelId}</div>
+                        )}
+                      </button>
+                    </li>
+                  );
+                })}
+              </div>
+              <div className={styles.extraChannel} onClick={props.onSourceInfoClick}>
+                {hasExtraChannels && (
+                  <LinkButton onClick={props.onSourceInfoClick}>
+                    +{channels.length - channelsToShow} {props.isConnected}
+                  </LinkButton>
+                )}
+              </div>
+            </div>
+          </div>
+
+          <div className={styles.channelButton}>
+            <button type="button" className={styles.addChannelButton} onClick={props.onAddChannelClick}>
+              <div className={styles.channelButtonIcon} title="Add a channel">
+                <PlusCircleIcon />
+              </div>
+            </button>
+          </div>
+        </>
+      )}
+    </>
+  );
+};
+export default ConnectedChannelsBySourceCard;
