+++ frontend/ui/src/pages/Channels/ConnectedChannelsList/index.module.scss
@@ -1,0 +1,162 @@
+@import 'assets/scss/fonts.scss';
+@import 'assets/scss/colors.scss';
+
+.wrapper {
+  background: white;
+  display: block;
+  border-radius: 10px;
+  padding-left: 96px;
+  padding-top: 88px;
+  width: 100%;
+  overflow-y: auto;
+  padding: 32px;
+  margin: 88px 1.5em 0 100px;
+  min-height: calc(100vh - 170px);
+}
+
+.headline {
+  @include font-xl;
+  font-weight: bold;
+  margin-bottom: 8px;
+}
+
+.backButton {
+  display: block;
+  margin-bottom: 16px;
+  cursor: pointer;
+  text-decoration: none;
+  max-width: 200px;
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.backIcon {
+  path {
+    fill: var(--color-airy-blue);
+  }
+  margin-right: 8px;
+}
+
+.inputContainer {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 32px;
+  width: 474px;
+  margin-top: 16px;
+
+  label {
+    margin-top: 24px;
+  }
+
+  input {
+    @include font-base;
+  }
+}
+
+.subtitle {
+  display: flex;
+  flex-direction: column;
+  @include font-s;
+  color: var(--color-airy-blue);
+  img {
+    width: 17px;
+    height: 13px;
+  }
+}
+
+.connectedChannels {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 32px;
+  width: 474px;
+  margin-top: 16px;
+}
+
+.channelRow {
+  display: flex;
+  flex-direction: row;
+}
+
+.connectedHint {
+  display: inline-flex;
+  align-items: center;
+  flex-direction: row;
+  border: 1px solid var(--color-soft-green);
+  padding: 0 8px;
+  border-radius: 4px;
+  margin-left: 8px;
+  font-family: 'Lato', sans-serif;
+  font-size: 0.8rem;
+  line-height: 1rem;
+  text-transform: uppercase;
+  color: var(--color-soft-green);
+}
+
+.headlineRow {
+  display: flex;
+  place-content: space-between;
+}
+
+.buttons {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+
+  button {
+    align-self: center;
+    height: 40px;
+    min-width: 40px;
+    margin-left: 16px;
+    background: var(--color-blue-white);
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+
+  button:hover {
+    background-color: var(--color-background-blue);
+  }
+
+  button:last-child {
+    background: var(--color-airy-blue);
+    svg {
+      path {
+        fill: white;
+      }
+    }
+  }
+  button:last-child:hover {
+    background-color: var(--color-airy-blue-hover);
+  }
+}
+
+.searchFieldButtons {
+  display: flex;
+  align-items: center;
+}
+
+.searchField {
+  min-width: 300px;
+  animation-name: searchFieldAnimation;
+  animation-duration: 500ms;
+  animation-fill-mode: forwards;
+}
+
+@keyframes searchFieldAnimation {
+  0% {
+    margin-left: 0px;
+    width: 0%;
+  }
+  100% {
+    margin-left: 16px;
+    width: 200px;
+  }
+}
+
+.noSearchMatch {
+  @include font-m;
+  font-weight: bold;
+  color: var(--color-text-contrast);
+  margin-bottom: 4px;
+}
