+++ frontend/ui/src/pages/Channels/Providers/Airy/ChatPlugin/ChatPluginConnect.module.scss
@@ -1,0 +1,155 @@
+@import 'assets/scss/fonts.scss';
+@import 'assets/scss/colors.scss';
+
+.wrapper {
+  background: white;
+  display: block;
+  border-radius: 10px;
+  padding-left: 96px;
+  padding-top: 88px;
+  width: 100%;
+  overflow-y: auto;
+  padding: 32px;
+  margin: 88px 1.5em 0 100px;
+  min-height: calc(100vh - 170px);
+}
+
+.headline {
+  @include font-xl;
+  font-weight: bold;
+  margin-bottom: 8px;
+}
+
+.backButton {
+  display: block;
+  margin-bottom: 16px;
+  cursor: pointer;
+  text-decoration: none;
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.backIcon {
+  path {
+    fill: var(--color-airy-blue);
+  }
+  margin-right: 8px;
+}
+
+.newPageParagraph {
+  margin: 24px 0;
+  color: var(--color-text-gray);
+}
+
+.updatePageParagraph {
+  margin: 24px 0;
+  color: var(--color-text-gray);
+}
+
+.formWrapper {
+  display: flex;
+}
+
+.formRow {
+  margin-bottom: 16px;
+}
+
+.settings {
+  width: 20rem;
+}
+
+.headline {
+  display: flex;
+}
+
+.headlineText {
+  flex-grow: 1;
+}
+
+.overview {
+  margin-top: 32px;
+}
+
+.listItem {
+  display: flex;
+  align-items: center;
+  margin-bottom: 8px;
+}
+
+.listChannelName {
+  flex-grow: 1;
+  border-bottom: 1px solid var(--color-light-gray);
+  height: 50px;
+  display: flex;
+  align-items: center;
+  padding-left: 16px;
+}
+
+.listButtons {
+  display: flex;
+  border-bottom: 1px solid var(--color-light-gray);
+  height: 50px;
+  align-items: center;
+}
+
+.listButtonEdit {
+  margin-right: 8px;
+  font-weight: normal;
+}
+
+.channelImage {
+  width: 40px;
+  height: 40px;
+  object-fit: cover;
+  border-radius: 100%;
+}
+
+.tabView {
+  display: flex;
+  margin-bottom: 16px;
+  list-style-type: none;
+
+  li {
+    margin-right: 16px;
+  }
+}
+
+.tabEntrySelected {
+  border-bottom: 2px solid var(--color-airy-blue);
+  a {
+    color: var(--color-text-contrast);
+    font-weight: bold;
+    text-decoration: none;
+  }
+}
+
+.tabEntry {
+  a {
+    color: var(--color-airy-blue);
+    text-decoration: none;
+  }
+}
+
+.installHint {
+  font-weight: bold;
+  margin-bottom: 8px;
+  code {
+    color: var(--color-red-alert);
+    background-color: var(--color-background-gray);
+    padding: 2px;
+  }
+}
+
+.codeArea {
+  @include font-s;
+  width: 40em;
+  height: 17em;
+  background-color: var(--color-background-gray);
+  color: var(--color-text-contrast);
+  border-radius: 8px;
+  border: 1px solid var(--color-dark-elements-gray);
+  resize: none;
+  padding: 8px;
+  margin-bottom: 16px;
+}
