+++ frontend/ui/src/pages/Channels/Providers/Facebook/Messenger/FacebookConnect.module.scss
@@ -1,0 +1,65 @@
+@import 'assets/scss/fonts.scss';
+@import 'assets/scss/colors.scss';
+
+.wrapper {
+  background: white;
+  display: block;
+  border-radius: 10px;
+  padding-left: 96px;
+  padding-top: 88px;
+  width: 100%;
+  overflow-y: auto;
+  padding: 32px;
+  margin: 88px 1.5em 0 100px;
+  min-height: calc(100vh - 170px);
+}
+
+.headline {
+  @include font-xl;
+  font-weight: bold;
+  margin-bottom: 8px;
+}
+
+.backButton {
+  display: block;
+  margin-bottom: 16px;
+  cursor: pointer;
+  text-decoration: none;
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+.backIcon {
+  path {
+    fill: var(--color-airy-blue);
+  }
+  margin-right: 8px;
+}
+
+.inputContainer {
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 32px;
+  width: 474px;
+  margin-top: 16px;
+
+  label {
+    margin-top: 24px;
+  }
+
+  input {
+    @include font-base;
+  }
+}
+
+.subtitle {
+  display: flex;
+  flex-direction: column;
+  @include font-s;
+  color: var(--color-airy-blue);
+  img {
+    width: 17px;
+    height: 13px;
+  }
+}
