--- frontend/ui/src/pages/Channels/Twilio/SmsWhatsappForm.module.scss
@@ -1,82 +1,0 @@
-@import '../../../../../assets/scss/fonts';
-@import '../../../../../assets/scss/colors';
-
-.formContainer {
-  margin-top: 24px;
-}
-
-.formContent {
-  display: block;
-  flex-direction: column;
-}
-
-.formContentNumber {
-  height: 62px;
-  width: 474px;
-  margin: 24px 0px 58px 0px;
-}
-
-.formImageCard {
-  display: flex;
-  box-sizing: border-box;
-  height: 100px;
-  width: 300px;
-  border: 1px solid var(--color-light-gray);
-  border-radius: 8px;
-  background-color: white;
-  margin-top: 24px;
-}
-
-.emptyLogo {
-  display: inline-flex;
-  height: 40px;
-  width: 40px;
-  margin: 16px 8px 42px 16px;
-  svg {
-    width: 40px;
-    height: 40px;
-  }
-}
-
-.imageUploadText {
-  display: inline-flex;
-  height: 24px;
-  width: 190px;
-  color: #212428;
-  font-size: 16px;
-  line-height: 24px;
-  margin: 16px 20px 2px 0;
-}
-
-.imageUploadSpecs {
-  display: inline-flex;
-  height: 48px;
-  width: 190px;
-  color: var(--color-text-gray);
-  font-size: 16px;
-  letter-spacing: 0;
-  line-height: 24px;
-}
-
-.imageUploadFooter {
-  height: 62px;
-  width: 474px;
-  margin: 24px 0px 24px 0px;
-  color: var(--color-text-gray);
-}
-
-.formContentName {
-  height: 62px;
-  width: 474px;
-  margin: 24px 0px 58px 0px;
-}
-
-.inputField {
-  @include font-base;
-  background-color: transparent;
-  width: 100%;
-  padding: 4px 8px;
-  border-radius: 4px;
-  border: 1px solid var(--color-dark-elements-gray);
-  margin: 5px 0px 5px 0px;
-}
