+++ frontend/ui/src/pages/Channels/index.tsx
@@ -1,18 +1,22 @@
 import React, {useEffect} from 'react';
 import _, {connect, ConnectedProps} from 'react-redux';
 import {Route, RouteComponentProps, Switch} from 'react-router-dom';
+
 import {listChannels} from '../../actions/channel';
 import {getClientConfig} from '../../actions/config';
 import {StateModel} from '../../reducers/index';
-import styles from './index.module.scss';
 import {allChannelsConnected} from '../../selectors/channels';
 import {setPageTitle} from '../../services/pageTitle';
-import ChannelsMainPage from './ChannelsMainPage';
-import FacebookConnect from './ChannelsMainPage/Sources/FacebookConnect';
-import ChannelsList from '../Channels/ConnectedChannelsList/ChannelsList';
-import ChatPluginConnect from './ChannelsMainPage/Sources/ChatPluginConnect';
-import TwilioSmsConnect from './ChannelsMainPage/Sources/TwilioSmsConnect';
-import TwilioWhatsappConnect from './ChannelsMainPage/Sources/TwilioWhatsappConnect';
+
+import MainPage from './MainPage';
+import FacebookConnect from './Providers/Facebook/Messenger/FacebookConnect';
+import ChatPluginConnect from './Providers/Airy/ChatPlugin/ChatPluginConnect';
+import ConnectedChannelsList from './ConnectedChannelsList';
+import TwilioSmsConnect from './Providers/Twilio/TwilioSmsConnect';
+import TwilioWhatsappConnect from './Providers/Twilio/TwilioWhatsappConnect';
+
+import styles from './index.module.scss';
+
 import {
   CHANNELS_TWILIO_SMS_ROUTE,
   CHANNELS_FACEBOOK_ROUTE,
@@ -46,7 +50,7 @@ const Channels = (props: ChannelsConnectProps) => {
 
   const renderChannels = () => (
     <div className={styles.channelsWrapper}>
-      <ChannelsMainPage channels={props.channels} config={props.config} />
+      <MainPage channels={props.channels} config={props.config} />
     </div>
   );
 
@@ -54,7 +58,7 @@ const Channels = (props: ChannelsConnectProps) => {
     <Switch>
       <Route path={[`${CHANNELS_FACEBOOK_ROUTE}/:channelId?`]} component={FacebookConnect} />
       <Route path={[`${CHANNELS_CHAT_PLUGIN_ROUTE}/:channelId?`]} component={ChatPluginConnect} />
-      <Route path={[`${CHANNELS_CONNECTED_ROUTE}/:source?`]} component={ChannelsList} />
+      <Route path={[`${CHANNELS_CONNECTED_ROUTE}/:source?`]} component={ConnectedChannelsList} />
       <Route path={[`${CHANNELS_TWILIO_SMS_ROUTE}/:channelId?`]} component={TwilioSmsConnect} />
       <Route path={[`${CHANNELS_TWILIO_WHATSAPP_ROUTE}/:channelId?`]} component={TwilioWhatsappConnect} />
       <Route path="/" render={renderChannels} />
